Initial publish for code backup
TigerGuard is not a new Discord bot, this bot has been in works slowly from ~2020 and slowly grew new features to meet then-current desires that the bot could handle at the time, getting to where it is now. This code as presented in this commit, is the whole embodiment of several rewrites for various segments of the program. Publishing for storing the code online and the usual other uses.