Debug logic and Rank updates
Improvements in the debug outputs to now reference a specific method that will now only print if the bot's main file for the debug boolean is true. Another improvement is moreso on the rank/levels xp output for vetting data. In the future this system will be fleshed out a bit more for server-agnositc usage, but still pending some things where the bot can eventually be used in DMs where the bot is not officially present, such as the case of group DMs or such.

Other news, minor patching of some errors in the button event class that appeared after API changes were introduced on the backend.