Debug mode update, config switching
Feature accidentally left out from previous update. This at last allows switching if the bot is in debug mode without it being hardcoded into the logic on startup.

To have debug messages show, the line 'debugMode = true' must exist in the config file. If the key/value does not exist or is false,  then debug messages will not be shown.