General backend update for level role logic
Changes here are intended to make level role logic a bit easier to work with in the future. Besides fixing some side effects, I've gone through ensuring that more proper sounding logic was used to get through the various things the changed code touches.  This is the first of a few changes to the leveling system, especially if I continue to mark things for a future guild-agnostic leveling system that the leveling system will support in tandem.